refactor GoodsController
