+++ src/main/java/com/exam/useful/adapter/goods/GoodsRequest.java
@@ -1,0 +1,2 @@
+package com.exam.useful.adapter.goods;public class GoodsRequest {
+}
