+++ src/main/java/com/exam/useful/domain/goods/model/Goods.java
