+++ src/main/java/com/exam/useful/domain/goods/model/GoodsPriceHistory.java
