+++ src/main/java/com/exam/useful/domain/goods/model/GoodsPriceHistoryKey.java
