+++ src/main/java/com/exam/useful/domain/goods/repository/GoodsRepository.java
