+++ src/main/java/com/exam/useful/usecase/goods/GoodsDTO.java
@@ -1,0 +1,2 @@
+package com.exam.useful.usecase.goods;public class GoodsDTO {
+}
