--- src/main/java/com/exam/useful/usecase/goods/GoodsService.java
@@ -1,8 +1,0 @@
-package com.exam.useful.usecase.goods;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface GoodsService {
-    String getMusinsaGoods(String goodsNo);
-}
